refactor: rename models path
