+++ src/main/java/com/coastalcare/models/enums/ClassificationPhoto.java
@@ -1,4 +1,4 @@
-package com.coastalcare.model.enums;
+package com.coastalcare.models.enums;
 
 public enum ClassificationPhoto {
     PLASTIC, OIL, MARINE_PLANTS, MARINE_LIFE,  ORGANIC_WASTE, CONSTRUCTION_DEBRIS, OTHER
