+++ src/main/java/com/coastalcare/models/enums/EventStatus.java
@@ -1,4 +1,4 @@
-package com.coastalcare.model.enums;
+package com.coastalcare.models.enums;
 
 public enum EventStatus {
     PLANNED, IN_PROGRESS, COMPLETED
