+++ src/main/java/com/coastalcare/models/enums/UserType.java
@@ -1,4 +1,4 @@
-package com.coastalcare.model.enums;
+package com.coastalcare.models.enums;
 
 public enum UserType {
     VOLUNTEER, CITIZEN
